backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.17
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
bcrypt==4.3.0
blinker==1.9.0
certifi==2025.1.31
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
deprecation==2.1.0
dill==0.4.0
Flask==3.1.0
flask-cors==5.0.1
frozenlist==1.6.0
gotrue==2.12.0
gunicorn==23.0.0
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.8
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mpmath==1.3.0
multidict==6.4.3
numpy==2.2.5
packaging==25.0
pbr==6.1.1
pillow==11.2.1
pluggy==1.5.0
postgrest==0.19.3
propcache==0.3.1
pydantic==2.11.3
pydantic_core==2.33.1
PyJWT==2.10.1
pytest==8.3.5
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
qiskit==2.0.0
realtime==2.4.2
reportlab==4.0.8
requests==2.32.3
rustworkx==0.16.0
scipy==1.15.2
setuptools==79.0.0
six==1.17.0
sniffio==1.3.1
stevedore==5.4.1
storage3==0.11.3
StrEnum==0.4.15
supabase==2.13.0
supafunc==0.9.4
SQLAlchemy
psycopg2-binary
symengine==0.13.0
sympy==1.13.3
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
websockets==14.2
Werkzeug==3.1.3
yarl==1.20.0
gevent
gunicorn
flask_socketio
gevent-websocket
Flask


